+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ##### Release date: 31/5/2022
 - Init project
 - improvement: add ant design
+- fix-bug: fix get posts error